requirements.txt
gradio
openai
openai-whisper
python-dotenv
transformers
torch